refactor: extract grid utility columns logic to separate mixins
